Finall Push
Organizer udah bisa waiting list, membuat export excell

+++ database/migrations/2026_03_30_151314_add_category_id_to_events_table.php
@@ -4,25 +4,25 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('events', function (Blueprint $table) {
-            $table->foreignId('category_id')->constrained()->onDelete('cascade');
-        });
+        // Cek dulu, kalau kolom BELUM ADA, baru tambahkan
+        if (!Schema::hasColumn('events', 'category_id')) {
+            Schema::table('events', function (Blueprint $table) {
+                $table->foreignId('category_id')->constrained()->onDelete('cascade');
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('events', function (Blueprint $table) {
-            //
+            $table->dropForeign(['category_id']);
+            $table->dropColumn('category_id');
         });
     }
 };
